Waypoint display zoom level improvements.
Improved `GME_displayPoints()` to make zooming behaviour more intuitive
when displaying waypoints.  On a listing page or when clicking through
to the main map, the map will be zoomed to display all the points, but
no further than level 15 when the points are close together.  This now
includes the primary point, even if GME doesn't draw a marker there.
When dragging'n'dropping, GME maintains the current zoom and just pans
to the centre of the dropped coordinates.