Move 'hide caches' from 'i' tool to widget
Removed the 'hide caches' function from the 'i' popup menu and added it
to the main GME widget, to make it easier to access.  Now you don't need
to click somewhere on the map before you can use it.  Also swapped to
using sprites for the widget icons, rather than embedding data: URIs.