Fix search by GC-code
Search by GC-code was broken under Greasemonkey 4 as it used a
GM.xmlHttpRequest which didn't pass on logged-in user credentials, and
so didn't return coordinates.  This version uses a native xmlHttpRequest
from the content script context.  This probably didn't work before while
the site was still using a mix of http and https, due to cross-site
issues.